+++ tests/lib/test_common.py
@@ -590,7 +590,17 @@ def test_validation_release_or_record_package(
         del validation_error_json["validator_value"]
         validation_error_jsons.append(validation_error_json)
 
-    assert validation_error_jsons == validation_error_jsons_expected
+    def strip_nones(list_of_dicts):
+        out = []
+        for a_dict in list_of_dicts:
+            out.append(
+                {key: value for key, value in a_dict.items() if value is not None}
+            )
+        return out
+
+    assert strip_nones(validation_error_jsons) == strip_nones(
+        validation_error_jsons_expected
+    )
 
 
 @pytest.mark.parametrize(